+++ src/Infrastructure/Legacy/Karo/KaroRequestParser.cs
@@ -56,6 +56,10 @@ public sealed class KaroRequestParser : IKaroRequestParser
             return value;
         }
 
-        return int.TryParse(value, out _) ? value : _encryption.Decrypt(value);
+        // Real Indici IDs (patientId/appointmentId) are `long`, not `int` - a plain int.TryParse
+        // silently overflows on real values above Int32.MaxValue (2,147,483,647), wrongly falling
+        // through to Decrypt() and returning an empty string. Confirmed real bug (2026-07-24), not a
+        // legacy quirk to preserve - fixed per Zohaib's direction ("in indici appointmentids type is long").
+        return long.TryParse(value, out _) ? value : _encryption.Decrypt(value);
     }
 }
